Wave Counter
The wave counter mentioned in a previous push has now been activated at a code level. It Currently displays the wave until the player eliminates an enemy.

The game also starts at either wave 2 or 3. Discussion is occurring as to whether this should be made a feature, to add to the difficulty or officially treated as a bug